Extract issue-related models to their own module
