+++ src/api/issues/list.rs
@@ -114,7 +114,7 @@ impl<'octo, 'b, 'c, 'd> ListIssuesBuilder<'octo, 'b, 'c, 'd> {
     }
 
     /// Sends the actual request.
-    pub async fn send(self) -> crate::Result<crate::Page<models::Issue>> {
+    pub async fn send(self) -> crate::Result<crate::Page<models::issues::Issue>> {
         let url = format!(
             "/repos/{owner}/{repo}/issues",
             owner = self.handler.owner,
