+++ src/api/issues/update.rs
@@ -71,7 +71,7 @@ impl<'octo, 'a, 'b, 'c, 'd, 'e> UpdateIssueBuilder<'octo, 'a, 'b, 'c, 'd, 'e> {
     }
 
     /// Send the actual request.
-    pub async fn send(self) -> Result<models::Issue> {
+    pub async fn send(self) -> Result<models::issues::Issue> {
         let route = format!(
             "/repos/{owner}/{repo}/issues/{issue}",
             owner = self.handler.owner,
